+++ internal/conn/errors_test.go
@@ -109,19 +109,19 @@ func TestIsBadConn(t *testing.T) {
 	}
 }
 
-func TestGrpcError(t *testing.T) {
-	err := withConnInfo(grpcStatus.Error(grpcCodes.Unavailable, "test"), 123, "test:123")
-	require.Equal(t, `rpc error: code = Unavailable desc = test`, err.Error())
+func TestWithConnInfo(t *testing.T) {
+	err := withConnInfo(grpcStatus.Error(grpcCodes.Unavailable, "test"), 100500, "example.com:2135")
+	require.ErrorIs(t, err, grpcStatus.Error(grpcCodes.Unavailable, "test"))
 	var nodeID interface {
 		NodeID() uint32
 	}
 	require.ErrorAs(t, err, &nodeID)
-	require.Equal(t, uint32(123), nodeID.NodeID())
+	require.Equal(t, uint32(100500), nodeID.NodeID())
 	var address interface {
 		Address() string
 	}
 	require.ErrorAs(t, err, &address)
-	require.Equal(t, "test:123", address.Address())
+	require.Equal(t, "example.com:2135", address.Address())
 	s, has := grpcStatus.FromError(err)
 	require.True(t, has)
 	require.Equal(t, grpcCodes.Unavailable, s.Code())