+++ internal/xerrors/operation_test.go
@@ -7,24 +7,30 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb_Issue"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xtest"
 )
 
 func TestIsOperationError(t *testing.T) {
 	for _, tt := range []struct {
+		name  string
 		err   error
 		codes []Ydb.StatusIds_StatusCode
 		match bool
 	}{
 		// check only operation error with any ydb status code
 		{
+			name:  xtest.CurrentFileLine(),
 			err:   &operationError{code: Ydb.StatusIds_BAD_REQUEST},
 			match: true,
 		},
 		{
+			name:  xtest.CurrentFileLine(),
 			err:   fmt.Errorf("wrapped: %w", &operationError{code: Ydb.StatusIds_BAD_REQUEST}),
 			match: true,
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err: Join(
 				fmt.Errorf("test"),
 				&operationError{code: Ydb.StatusIds_BAD_REQUEST},
@@ -34,16 +40,19 @@ func TestIsOperationError(t *testing.T) {
 		},
 		// match ydb status code
 		{
+			name:  xtest.CurrentFileLine(),
 			err:   &operationError{code: Ydb.StatusIds_BAD_REQUEST},
 			codes: []Ydb.StatusIds_StatusCode{Ydb.StatusIds_BAD_REQUEST},
 			match: true,
 		},
 		{
+			name:  xtest.CurrentFileLine(),
 			err:   fmt.Errorf("wrapped: %w", &operationError{code: Ydb.StatusIds_BAD_REQUEST}),
 			codes: []Ydb.StatusIds_StatusCode{Ydb.StatusIds_BAD_REQUEST},
 			match: true,
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err: Join(
 				fmt.Errorf("test"),
 				&operationError{code: Ydb.StatusIds_BAD_REQUEST},
@@ -54,16 +63,19 @@ func TestIsOperationError(t *testing.T) {
 		},
 		// no match ydb status code
 		{
+			name:  xtest.CurrentFileLine(),
 			err:   &operationError{code: Ydb.StatusIds_BAD_REQUEST},
 			codes: []Ydb.StatusIds_StatusCode{Ydb.StatusIds_ABORTED},
 			match: false,
 		},
 		{
+			name:  xtest.CurrentFileLine(),
 			err:   fmt.Errorf("wrapped: %w", &operationError{code: Ydb.StatusIds_BAD_REQUEST}),
 			codes: []Ydb.StatusIds_StatusCode{Ydb.StatusIds_ABORTED},
 			match: false,
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err: Join(
 				fmt.Errorf("test"),
 				&operationError{code: Ydb.StatusIds_BAD_REQUEST},
@@ -73,7 +85,7 @@ func TestIsOperationError(t *testing.T) {
 			match: false,
 		},
 	} {
-		t.Run("", func(t *testing.T) {
+		t.Run(tt.name, func(t *testing.T) {
 			require.Equal(t, tt.match, IsOperationError(tt.err, tt.codes...))
 		})
 	}
@@ -81,10 +93,12 @@ func TestIsOperationError(t *testing.T) {
 
 func TestIsOperationErrorTransactionLocksInvalidated(t *testing.T) {
 	for _, tt := range [...]struct {
+		name  string
 		err   error
 		isTLI bool
 	}{
 		{
+			name: xtest.CurrentFileLine(),
 			err: Operation(
 				WithStatusCode(Ydb.StatusIds_ABORTED),
 				WithIssues([]*Ydb_Issue.IssueMessage{{
@@ -94,6 +108,7 @@ func TestIsOperationErrorTransactionLocksInvalidated(t *testing.T) {
 			isTLI: true,
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err: Operation(
 				WithStatusCode(Ydb.StatusIds_OVERLOADED),
 				WithIssues([]*Ydb_Issue.IssueMessage{{
@@ -103,12 +118,14 @@ func TestIsOperationErrorTransactionLocksInvalidated(t *testing.T) {
 			isTLI: false,
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err: Operation(
 				WithStatusCode(Ydb.StatusIds_ABORTED),
 			),
 			isTLI: false,
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err: Operation(
 				WithStatusCode(Ydb.StatusIds_ABORTED),
 				WithIssues([]*Ydb_Issue.IssueMessage{{
@@ -120,7 +137,7 @@ func TestIsOperationErrorTransactionLocksInvalidated(t *testing.T) {
 			isTLI: true,
 		},
 	} {
-		t.Run("", func(t *testing.T) {
+		t.Run(tt.name, func(t *testing.T) {
 			require.Equal(t, tt.isTLI, IsOperationErrorTransactionLocksInvalidated(tt.err))
 		})
 	}
@@ -128,22 +145,32 @@ func TestIsOperationErrorTransactionLocksInvalidated(t *testing.T) {
 
 func Test_operationError_Error(t *testing.T) {
 	for _, tt := range []struct {
+		name string
 		err  error
 		text string
 	}{
 		{
+			name: xtest.CurrentFileLine(),
 			err:  Operation(WithStatusCode(Ydb.StatusIds_BAD_REQUEST), WithAddress("localhost")),
 			text: "operation/BAD_REQUEST (code = 400010, address = localhost)",
 		},
 		{
+			name: xtest.CurrentFileLine(),
+			err:  Operation(WithStatusCode(Ydb.StatusIds_BAD_REQUEST), WithNodeID(100500)),
+			text: "operation/BAD_REQUEST (code = 400010, nodeID = 100500)",
+		},
+		{
+			name: xtest.CurrentFileLine(),
 			err:  Operation(WithStatusCode(Ydb.StatusIds_BAD_REQUEST)),
 			text: "operation/BAD_REQUEST (code = 400010)",
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err:  Operation(WithStatusCode(Ydb.StatusIds_BAD_SESSION)),
 			text: "operation/BAD_SESSION (code = 400100)",
 		},
 		{
+			name: xtest.CurrentFileLine(),
 			err: Operation(WithStatusCode(Ydb.StatusIds_PRECONDITION_FAILED), WithIssues([]*Ydb_Issue.IssueMessage{
 				{
 					Message:   "issue one",
@@ -177,7 +204,7 @@ func Test_operationError_Error(t *testing.T) {
 			text: "operation/PRECONDITION_FAILED (code = 400120, issues = [{15:3 => #1 'issue one'},{#2 'issue two' [{test.yql:16:4 => #3 'issue three'},{#4 'issue four'}]}])", //nolint:lll
 		},
 	} {
-		t.Run("", func(t *testing.T) {
+		t.Run(tt.name, func(t *testing.T) {
 			require.Equal(t, tt.text, tt.err.Error())
 		})
 	}